+++ My-eCommerce.sln
@@ -62,8 +62,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CatalogManagement.Applicati
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CatalogManagement.Domain", "src\CatalogManagement\CatalogManagement.Domain\CatalogManagement.Domain.csproj", "{1CD8E5A1-1AE3-4651-ABD4-DD0356C6B283}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ORM.Migrations", "src\ORM.Migrations\ORM.Migrations.csproj", "{A3BD7429-F35F-4385-B89A-42A2B61A76A6}"
-EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CatalogManagement.Unit", "tests\CatalogManagement.Unit\CatalogManagement.Unit.csproj", "{F4B4EED8-BBB6-4DDF-9149-189BCF5A757C}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Common", "Common", "{A3B098A0-8001-1A2B-7DF0-C3B37F63E4C6}"
@@ -87,6 +85,8 @@ EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CatalogManagement.Functionl", "tests\CatalogManagement.Functionl\CatalogManagement.Functionl.csproj", "{D20C444E-6337-42A3-A5A2-88CB575108AB}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SalesManagement.Functionl", "tests\SalesManagement.Functionl\SalesManagement.Functionl.csproj", "{8DD7839B-816A-4D36-A761-6A784678D8F8}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ORM.Migrations", "src\ORM.Migrations\ORM.Migrations.csproj", "{6CF5EB6E-BA53-4182-8D6A-E43AB1E8C7FB}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -262,18 +262,6 @@ Global
 		{1CD8E5A1-1AE3-4651-ABD4-DD0356C6B283}.Release|x64.Build.0 = Release|Any CPU
 		{1CD8E5A1-1AE3-4651-ABD4-DD0356C6B283}.Release|x86.ActiveCfg = Release|Any CPU
 		{1CD8E5A1-1AE3-4651-ABD4-DD0356C6B283}.Release|x86.Build.0 = Release|Any CPU
-		{A3BD7429-F35F-4385-B89A-42A2B61A76A6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A3BD7429-F35F-4385-B89A-42A2B61A76A6}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A3BD7429-F35F-4385-B89A-42A2B61A76A6}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{A3BD7429-F35F-4385-B89A-42A2B61A76A6}.Debug|x64.Build.0 = Debug|Any CPU
-		{A3BD7429-F35F-4385-B89A-42A2B61A76A6}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{A3BD7429-F35F-4385-B89A-42A2B61A76A6}.Debug|x86.Build.0 = Debug|Any CPU
-		{A3BD7429-F35F-4385-B89A-42A2B61A76A6}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A3BD7429-F35F-4385-B89A-42A2B61A76A6}.Release|Any CPU.Build.0 = Release|Any CPU
-		{A3BD7429-F35F-4385-B89A-42A2B61A76A6}.Release|x64.ActiveCfg = Release|Any CPU
-		{A3BD7429-F35F-4385-B89A-42A2B61A76A6}.Release|x64.Build.0 = Release|Any CPU
-		{A3BD7429-F35F-4385-B89A-42A2B61A76A6}.Release|x86.ActiveCfg = Release|Any CPU
-		{A3BD7429-F35F-4385-B89A-42A2B61A76A6}.Release|x86.Build.0 = Release|Any CPU
 		{F4B4EED8-BBB6-4DDF-9149-189BCF5A757C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{F4B4EED8-BBB6-4DDF-9149-189BCF5A757C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{F4B4EED8-BBB6-4DDF-9149-189BCF5A757C}.Debug|x64.ActiveCfg = Debug|Any CPU
@@ -406,6 +394,18 @@ Global
 		{8DD7839B-816A-4D36-A761-6A784678D8F8}.Release|x64.Build.0 = Release|Any CPU
 		{8DD7839B-816A-4D36-A761-6A784678D8F8}.Release|x86.ActiveCfg = Release|Any CPU
 		{8DD7839B-816A-4D36-A761-6A784678D8F8}.Release|x86.Build.0 = Release|Any CPU
+		{6CF5EB6E-BA53-4182-8D6A-E43AB1E8C7FB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{6CF5EB6E-BA53-4182-8D6A-E43AB1E8C7FB}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{6CF5EB6E-BA53-4182-8D6A-E43AB1E8C7FB}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{6CF5EB6E-BA53-4182-8D6A-E43AB1E8C7FB}.Debug|x64.Build.0 = Debug|Any CPU
+		{6CF5EB6E-BA53-4182-8D6A-E43AB1E8C7FB}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{6CF5EB6E-BA53-4182-8D6A-E43AB1E8C7FB}.Debug|x86.Build.0 = Debug|Any CPU
+		{6CF5EB6E-BA53-4182-8D6A-E43AB1E8C7FB}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{6CF5EB6E-BA53-4182-8D6A-E43AB1E8C7FB}.Release|Any CPU.Build.0 = Release|Any CPU
+		{6CF5EB6E-BA53-4182-8D6A-E43AB1E8C7FB}.Release|x64.ActiveCfg = Release|Any CPU
+		{6CF5EB6E-BA53-4182-8D6A-E43AB1E8C7FB}.Release|x64.Build.0 = Release|Any CPU
+		{6CF5EB6E-BA53-4182-8D6A-E43AB1E8C7FB}.Release|x86.ActiveCfg = Release|Any CPU
+		{6CF5EB6E-BA53-4182-8D6A-E43AB1E8C7FB}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -433,7 +433,6 @@ Global
 		{78F1AFEC-AFDF-46D4-A81B-FE0BEEE93729} = {BD3812F8-8783-42C6-B231-8A032EF9481F}
 		{18EBE193-A874-4A84-9D42-7754C0CE1EDC} = {7DC3E4B1-675C-4122-AA58-4BCBA06C4A82}
 		{1CD8E5A1-1AE3-4651-ABD4-DD0356C6B283} = {77C73D0D-E122-4C65-B63B-B89C8E28F69E}
-		{A3BD7429-F35F-4385-B89A-42A2B61A76A6} = {BD3812F8-8783-42C6-B231-8A032EF9481F}
 		{F4B4EED8-BBB6-4DDF-9149-189BCF5A757C} = {9057D64B-F3A3-4127-89EA-622C2430FD25}
 		{A3B098A0-8001-1A2B-7DF0-C3B37F63E4C6} = {1ADFD29B-9AE9-4E79-923B-072E807A2AAF}
 		{C1A80877-EE45-425D-B2F6-22CA95104E44} = {A3B098A0-8001-1A2B-7DF0-C3B37F63E4C6}
@@ -446,6 +445,7 @@ Global
 		{DF4D4F0B-AD1B-4077-8BE8-CDF1FEBDC02B} = {D641472C-1E35-4F9D-8949-90E6C17D7078}
 		{D20C444E-6337-42A3-A5A2-88CB575108AB} = {95501476-3BCD-48DA-A4E5-696A0B78A0BD}
 		{8DD7839B-816A-4D36-A761-6A784678D8F8} = {95501476-3BCD-48DA-A4E5-696A0B78A0BD}
+		{6CF5EB6E-BA53-4182-8D6A-E43AB1E8C7FB} = {040F270A-3788-4E88-99A6-F236286DE3E7}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {13EAB81A-C6B2-4C22-B4DE-C6603FC445B6}
