+++ src/SalesManagement/SalesManagement.Application/Carts/CreateCart/CreateCartHandler.cs
@@ -6,6 +6,7 @@ using SalesManagement.Application.Repositories;
 using SalesManagement.Application.Services;
 using SalesManagement.Domain.Entities;
 using SalesManagement.Domain.Specifications;
+using SalesManagement.Domain.ValueObjects;
 
 namespace SalesManagement.Application.Carts.CreateCart;
