+++ src/SalesManagement/SalesManagement.Domain/Entities/SaleItem.cs
@@ -1,5 +1,6 @@
 using Common.DomainCommon;
 using SalesManagement.Domain.Common;
+using SalesManagement.Domain.ValueObjects;
 
 namespace SalesManagement.Domain.Entities;
