+++ tests/SalesManagement.Integration/Sales/TestData/CartFeatureTestData.cs
@@ -3,6 +3,7 @@ using CatalogManagement.Domain.Entities.ValueObjects;
 using SalesManagement.Application.Services.DTOs;
 using SalesManagement.Domain.Entities;
 using SalesManagement.Domain.Enums;
+using SalesManagement.Domain.ValueObjects;
 
 namespace SalesManagement.Integration.Sales.TestData;
